feat: 修改文档功能由swagger迁到springdoc feat: 统一异常处理抽取到公公模块 fix: 修复版本更新带来的@FeignClient和@RequestMapping同时出现在类上引发的异常问题

+++ gulimall-common/src/main/java/com/laughingather/gulimall/common/entity/query/PageQuery.java
@@ -15,11 +15,11 @@ public class PageQuery {
     /**
      * 设置默认分页为第一页
      */
-    private Integer pn = 1;
+    private Long pn = 1L;
 
     /**
      * 设置默认每页条数为十条
      */
-    private Integer ps = 10;
+    private Long ps = 10L;
 
 }
